rename project amypad => amypet
- fixes #14
