+++ .github/workflows/test.yml
@@ -94,7 +94,7 @@ jobs:
       env:
         GITHUB_TOKEN: ${{ secrets.GH_TOKEN }}
       with:
-        name: amypad ${{ steps.meta.outputs.tag }} alpha
+        name: amypet ${{ steps.meta.outputs.tag }} alpha
         body_path: _CHANGES.md
         draft: true
         files: |
