+++ .gitignore
@@ -2,7 +2,7 @@
 __pycache__/
 
 # build
-/amypad/_dist_ver.py
+/amypet/_dist_ver.py
 /build/
 /dist/
 /*.egg*/
