+++ amypet/centiloid/f_3Segment.m
