+++ amypet/centiloid/f_noNaN.m
