+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name=amypad
+name=amypet
 description=Amyloid imaging to prevent Alzheimer's Disease
 long_description=file: README.rst
 long_description_content_type=text/x-rst
@@ -7,14 +7,14 @@ license=MPL 2.0
 license_file=LICENCE.md
 url=https://amypad.eu
 project_urls=
-    Repository=https://github.com/AMYPAD/amypad
-    Changelog=https://github.com/AMYPAD/amypad/releases
-    Documentation=https://github.com/AMYPAD/amypad/#amypad
+    Repository=https://github.com/AMYPAD/amypet
+    Changelog=https://github.com/AMYPAD/amypet/releases
+    Documentation=https://github.com/AMYPAD/amypet/#amypet
 maintainer=Casper da Costa-Luis
 maintainer_email=casper.dcl@physics.org
 keywords=pet, alzheimers
 platforms=any
-provides=amypad
+provides=amypet
 classifiers=
     Development Status :: 3 - Alpha
     Intended Audience :: Developers
@@ -61,8 +61,8 @@ gui=Gooey>=1.0.8
 niftypet=niftypet>=0.0.1
 [options.entry_points]
 console_scripts=
-    amypad=amypad.cli:main
-    amypad.gui=amypad.gui:main
+    amypet=amypet.cli:main
+    amypet.gui=amypet.gui:main
 [options.packages.find]
 exclude=tests
 [options.package_data]
@@ -87,12 +87,12 @@ split_before_closing_bracket=False
 [isort]
 profile=black
 line_length=99
-known_first_party=amypad,tests
+known_first_party=amypet,tests
 
 [tool:pytest]
 timeout=10
 log_level=INFO
 python_files=tests/test_*.py
-addopts=-v --tb=short -rxs -W=error -n=auto --durations=0 --durations-min=1 --cov=amypad --cov-report=term-missing --cov-report=xml
+addopts=-v --tb=short -rxs -W=error -n=auto --durations=0 --durations-min=1 --cov=amypet --cov-report=term-missing --cov-report=xml
 filterwarnings=
     ignore:numpy.ufunc size changed.*:RuntimeWarning
